Wrap calls to _delete_dataset_if_exists in catch_errors
This commit wraps the calls to _delete_dataset_if_exists with
catch_errors. This causes the GDAL error messages to be silenced. Since
calling _delete_dataset_if_exists causes GDAL to display an `ERROR
4....` message, and the act is deliberate to force a new file to be
created, it makes sense to ignore this error message, as it is not an
actual error.